+++ packages/calculator-desmos/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @pie-players/pie-calculator-desmos
 
+## 0.3.25
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-calculator@0.3.25
+
+## 0.3.25
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-calculator@0.3.25
+
+## 0.3.24
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-calculator@0.3.24
+
 ## 0.3.23
 
 ### Patch Changes