+++ packages/section-player-tools-instrumentation-debugger/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @pie-players/pie-section-player-tools-instrumentation-debugger
 
+## 0.3.25
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-players-shared@0.3.25
+  - @pie-players/pie-section-player-tools-shared@0.3.25
+  - @pie-players/pie-theme@0.3.25
+
+## 0.3.25
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-players-shared@0.3.25
+  - @pie-players/pie-section-player-tools-shared@0.3.25
+  - @pie-players/pie-theme@0.3.25
+
+## 0.3.24
+
+### Patch Changes
+
+- Temporary release changeset: patch all publishable packages to keep lockstep versions.
+- Updated dependencies
+  - @pie-players/pie-players-shared@0.3.24
+  - @pie-players/pie-section-player-tools-shared@0.3.24
+  - @pie-players/pie-theme@0.3.24
+
 ## 0.3.23
 
 ### Patch Changes